+++ roda-core/roda-core/src/main/java/org/roda/core/plugins/plugins/ingest/characterization/JpylyzerPlugin.java
@@ -36,11 +36,13 @@ import org.roda.core.data.exceptions.RODAException;
 import org.roda.core.data.v2.ip.AIP;
 import org.roda.core.data.v2.ip.File;
 import org.roda.core.data.v2.ip.Representation;
+import org.roda.core.data.v2.ip.StoragePath;
 import org.roda.core.data.v2.jobs.PluginParameter;
 import org.roda.core.data.v2.jobs.PluginType;
 import org.roda.core.data.v2.jobs.Report;
 import org.roda.core.index.IndexService;
 import org.roda.core.model.ModelService;
+import org.roda.core.model.utils.ModelUtils;
 import org.roda.core.plugins.Plugin;
 import org.roda.core.plugins.PluginException;
 import org.roda.core.storage.Binary;
@@ -104,20 +106,23 @@ public class JpylyzerPlugin implements Plugin<AIP> {
       for (String representationID : aip.getRepresentationIds()) {
         LOGGER.debug("Processing representation " + representationID + " from AIP " + aip.getId());
         try {
-          Representation representation = model.retrieveRepresentation(aip.getId(), representationID);
-          for (String fileID : representation.getFileIds()) {
-            try {
-              LOGGER.debug("Processing file " + fileID + " from " + representationID + " of AIP " + aip.getId());
-              File file = model.retrieveFile(aip.getId(), representationID, fileID);
-              Binary binary = storage.getBinary(file.getStoragePath());
+          Iterable<File> allFiles = model.listAllFiles(aip.getId(), representationID);
+          for (File file : allFiles) {
+            if (!file.isDirectory()) {
+              // TODO check if file is JPEG2000
+              try {
+                LOGGER.debug("Processing file: " + file);
+                StoragePath storagePath = ModelUtils.getRepresentationFilePath(file);
+                Binary binary = storage.getBinary(storagePath);
 
-              Path ffProbeResults = FFProbePluginUtils.runFFProbe(file, binary, getParameterValues());
-              Binary resource = (Binary) FSUtils.convertPathToResource(ffProbeResults.getParent(), ffProbeResults);
-              model.createOtherMetadata(aip.getId(), representationID, file.getStoragePath().getName() + ".xml",
-                "jpylyzer", resource);
-              ffProbeResults.toFile().delete();
-            } catch (RODAException | IOException e) {
-              LOGGER.error("Error processing AIP " + aip.getId() + ": " + e.getMessage());
+                Path ffProbeResults = JpylyzerPluginUtils.runJpylyzer(file, binary, getParameterValues());
+
+                Binary resource = (Binary) FSUtils.convertPathToResource(ffProbeResults.getParent(), ffProbeResults);
+                model.createOtherMetadata(aip.getId(), representationID, file.getId() + ".xml", "jpylyzer", resource);
+                ffProbeResults.toFile().delete();
+              } catch (RODAException | IOException sse) {
+                LOGGER.error("Error processing AIP " + aip.getId() + ": " + sse.getMessage());
+              }
             }
           }
         } catch (RODAException mse) {
